branchout-llm-service/requirements.txt
fastmcp
pydantic
httpx
uvicorn
python-dotenv